Introduce a prior loss function
